refactor(gui-v2): split sidebar into smaller components
